Modified nvmemory subclass.
Changed the init_mem function of the node subclass to allow for directly
creating the structures of implementing classes so their constructor is
correctly called. Fixed classes that use this to use it the new way.

Also did some minor changes to fstring and history classes along with
changing init_mem.