Modified file class.
Changed the file class so it does not inherit the nvmemory node class and
instead adds the header data in as a seperate class. This makes the
classes more flat and allows for the enduser of the library to use
nvmemory node as an inherited class for their data plugin classes.