Modified multiple classes.
Changed constructor of cldevice to private since only cldevlist should
be using it.

As a result, cldevlist needed to be modified so it was the only class
using cldevice's constructor. Changed double list to hold a unique
pointer reference to cldevice objects so cldevlist itself called the
constructor instead of using emplace_back that forwarded its use.