Worked on analytic system.
Added new setupanalyticdialog class that will be used for setting
arguments of an analytic through a dialog GUI. As a result had to modify
the mainwindow class to be a singleton class so the console can be
accessed from anywhere in the program.

Also added additional virtual functions for the abstractanalytic class
that will be needed for argument processing.