Made additions to core module.
Connected metadata system with data objects. Changed metadata class to
emetadata class that is now part of the public API so data objects and
add their own metadata. Added to abstract analytic class to analytics
can now open new data objects on the fly as they are executing. Added
base opencl class to eopenclevent so it does not throw exceptions.