Worked on entire program.
Continued working of setup analytic dialog GUI and as a result had to
work on mostly the core module again to add more functinoality to the
abstract analytic class. Said analytic class is now the manager of all
input/output files/data objects it now uses to make it easier for the gui
and console to handle the entire process.