Added to entire project.
Added many helper methods to various classes to help more with MPI and
opencl. Exceptions thrown in an opencl thread are now caught and
rethrown in the main thread. Blocks are handled much better in regards
to exracting their index.

Along with other minor changes to make it all work better.