Added to analytic run MPI system.
Fixed bug that caused the master node to steal a GPU and not use it.
This required a major expansion of the QMPI class along with major
modification of the analytic run MPI classes. As a result added a new
base MPI class for analytic runs that both master and slave inherits
from.

Also made slight modification to analytic run chunk class in regards to
runner initialization.