Added new settings dialog.
Finished adding code for settings dialog to edit all the new settings
for this new version of ACE. This was the last part of the GUI library
that needed refactoring. The entire GUI library is not refactored and
can compile without errors. This now makes it so the entire ACE library,
minus unit tests, can compile without error.

The GUI library is still unfinished; it requires testing and addition of
full comments to all refactored classes.