Merged GUI library.
Heavily modified example program to work for both CLI and GUI. Did this
by making the majority of its code into a static library and the CLI or
GUI version of the example program using that static library. Now both
the CLI binary aceex and the GUI binary qaceex are compiled and
installed.