Refactored code and comments.
Moved the CUDA run thread class outside of its parent class to be its
own root class because it is a qt object. Updated the code and comments
for the CUDA run class and its now separate CUDA run thread class.

Fixed some minor grammar errors in the comments of OpenCL run and run
thread classses.