Improved CUDA helper classes.
Removed ALL unnecesary macro methods that Ben used for error checking in
all of his CUDA code. This was removed because using macro functions
unless it is impossible to do otherwise is very bad practice. In this
case those macros were not needed and simply served at best a minor
convenience to the programmer writing the code.