Added to edebug and chunkrun.
Added additional define check for the EDEBUG_FUNC macro that requires the
EDEBUG define for it to actually do anything. EDEBUG must be defined at
the BEGINNING of a source file BEFORE any include statements. This allows
for whitelisting only specific source files. EDEBUG should NEVER be defined
in a header file.

Added to the chunk run so it emits progress changed signals.