Fixed MPI bug.
Fixed a very special case bug with MPI that would only occur if the user
set the number of processes to be 1. While this seems silly it is still
a bug. This happened because ACE cannot differentiate between an MPI run
of one and running the program directly. In either case a Single
analytic run is started and this never calls the shutdown method. To
make sure MPI is properly shutdown it was added in the cli library at
the very end of program execution.