Fixed QMPI shutdown bug.
The other bug fix caused another hidden bug to crop up. The shutdown
method never resets the instance pointer to null, now it does. This now
makes it safe to call shutdown any number of times, and it will only
actually shutdown MPI once the first time it is called.